@@ -20,7 +20,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,15 +34,15 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  google_fonts: ^1.1.2
-  carousel_slider: ^3.0.0
-  cached_network_image: ^2.3.2
+  cupertino_icons: ^1.0.2
+  google_fonts: ^4.0.3
+  carousel_slider: ^4.2.1
+  cached_network_image: ^3.2.0
   lottie: ^2.2.0
   flutter_datetime_picker: ^1.5.1
   material_dialogs: ^1.1.4
-  provider: ^4.3.3
-  http: ^0.12.2
+  provider: ^6.0.1
+  http: ^0.13.3
   shimmer: ^3.0.0
   shared_preferences: ^2.0.9
   intl: ^0.18.1
