@@ -2,15 +2,15 @@ import 'package:pedikia/models/category_model.dart';
 import 'package:pedikia/models/gallery_model.dart';
 
 class ServiceModel {
-  int id;
-  String name;
-  double price;
-  String description;
-  String tags;
-  CategoryModel category;
-  DateTime createdAt;
-  DateTime updatedAt;
-  List<GalleryModel> galleries;
+  late int? id;
+  late String? name;
+  late double? price;
+  late String? description;
+  late String? tags;
+  late CategoryModel? category;
+  late DateTime? createdAt;
+  late DateTime? updatedAt;
+  late List<GalleryModel>? galleries;
 
   ServiceModel({
     this.id,
@@ -45,8 +45,8 @@ class ServiceModel {
       'price': price,
       'description': description,
       'tags': tags,
-      'category': category.toJson(),
-      'galleries': galleries.map((gallery) => gallery.toJson()).toList(),
+      'category': category!.toJson(),
+      'galleries': galleries!.map((gallery) => gallery.toJson()).toList(),
       'created_at': createdAt.toString(),
       'updated_at': updatedAt.toString(),
     };
