@@ -5,23 +5,23 @@ import 'package:http/http.dart' as http;
 import 'package:shared_preferences/shared_preferences.dart';
 
 class OrderService {
-  // String baseUrl = "https://testing.tanpabatasgroup.com/api";
-  String baseUrl = "http://10.0.2.2:8000/api";
+  String baseUrl = "https://testing.tanpabatasgroup.com/public/api";
+  // String baseUrl = "http://10.0.2.2:8000/api";
 
   Future<List<OrderModel>> getOrder(
-      // String token,
-      ) async {
+    String token,
+  ) async {
     var url = "$baseUrl/transactions";
 
-    final sharedPreferences = await SharedPreferences.getInstance();
-    final token = sharedPreferences.getString('token');
+    // final sharedPreferences = await SharedPreferences.getInstance();
+    // final token = sharedPreferences.getString('token');
     var headers = {
       "Content-Type": "application/json",
       'Authorization': token,
     };
 
     var response = await http.get(
-      url,
+      Uri.parse(url),
       headers: headers,
     );
 
@@ -30,12 +30,10 @@ class OrderService {
       List data = jsonDecode(response.body)['data']['data'];
       List<OrderModel> orders = [];
       // print(response.body);
-      if (data != null) {
-        for (var item in data) {
-          // print(item);
-          // print(data);
-          orders.add(OrderModel.fromJson(item));
-        }
+      for (var item in data) {
+        // print(item);
+        // print(data);
+        orders.add(OrderModel.fromJson(item));
       }
 
       // return orders;
